feat(api): add validators api
