+++ consumer/consumer.go
@@ -6,9 +6,9 @@ import (
 	"encoding/hex"
 	"encoding/json"
 	"fmt"
+	"net/http"
 	"os"
 	"os/signal"
-	"strconv"
 	"syscall"
 	"time"
 
@@ -17,7 +17,6 @@ import (
 	"github.com/chainbound/valtrack/log"
 	"github.com/chainbound/valtrack/types"
 	_ "github.com/mattn/go-sqlite3"
-	ma "github.com/multiformats/go-multiaddr"
 	"github.com/nats-io/nats.go"
 	"github.com/nats-io/nats.go/jetstream"
 	"github.com/rs/zerolog"
@@ -43,28 +42,6 @@ type Consumer struct {
 
 	chClient *ch.ClickhouseClient
 	db       *sql.DB
-}
-
-func setupDatabase(db *sql.DB) error {
-	createTableQuery := `
-    CREATE TABLE IF NOT EXISTS validator_tracker (
-        PeerID TEXT PRIMARY KEY,
-        ENR TEXT,
-        Multiaddr TEXT,
-        IP TEXT,
-        Port INTEGER,
-        LastSeen TEXT,
-        LastEpoch INTEGER,
-        PossibleValidator BOOLEAN,
-        AverageValidatorCount INTEGER,
-        NumObservations INTEGER
-    );
-    `
-	_, err := db.Exec(createTableQuery)
-	if err != nil {
-		return err
-	}
-	return nil
 }
 
 func RunConsumer(cfg *ConsumerConfig) {
@@ -189,6 +166,13 @@ func RunConsumer(cfg *ConsumerConfig) {
 	}()
 
 	go consumer.HandleValidatorMetadataEvent()
+
+	// Start the HTTP server
+	http.HandleFunc("/validators", createGetValidatorsHandler(db))
+
+	if err := http.ListenAndServe(":8080", nil); err != nil {
+		log.Fatal().Err(err).Msg("Error starting HTTP server")
+	}
 
 	// Gracefully shutdown
 	quit := make(chan os.Signal, 1)
@@ -331,90 +315,5 @@ func (c *Consumer) storeMetadataReceivedEvent(event types.MetadataReceivedEvent)
 		c.log.Err(err).Msg("Failed to write metadata_received event to Parquet file")
 	} else {
 		c.log.Trace().Msg("Wrote metadata_received event to Parquet file")
-	}
-}
-
-func (c *Consumer) HandleValidatorMetadataEvent() error {
-	for {
-		select {
-		case event := <-c.validatorMetadataChan:
-			c.log.Trace().Any("event", event).Msg("Received validator event")
-
-			maddr, err := ma.NewMultiaddr(event.Multiaddr)
-			if err != nil {
-				c.log.Error().Err(err).Msg("Invalid multiaddr")
-				continue
-			}
-
-			ip, err := maddr.ValueForProtocol(ma.P_IP4)
-			if err != nil {
-				ip, err = maddr.ValueForProtocol(ma.P_IP6)
-				if err != nil {
-					c.log.Error().Err(err).Msg("Invalid IP in multiaddr")
-					continue
-				}
-			}
-
-			portStr, err := maddr.ValueForProtocol(ma.P_TCP)
-			if err != nil {
-				c.log.Error().Err(err).Msg("Invalid port in multiaddr")
-				continue
-			}
-
-			port, err := strconv.Atoi(portStr)
-			if err != nil {
-				c.log.Error().Err(err).Msg("Invalid port number")
-				continue
-			}
-
-			isValidator := true
-			longLived := indexesFromBitfield(event.MetaData.Attnets)
-			shortLived := extractShortLivedSubnets(event.SubscribedSubnets, longLived)
-			// If there are no short lived subnets, then the peer is not a validator
-			if len(shortLived) == 0 {
-				isValidator = false
-			}
-
-			prevNumObservations := uint64(0)
-			prevAvgValidatorCount := int32(0)
-			err = c.db.QueryRow("SELECT NumObservations, AverageValidatorCount FROM validator_tracker WHERE PeerID = ?", event.ID).Scan(&prevNumObservations, &prevAvgValidatorCount)
-
-			currValidatorCount := 1 + (len(shortLived)-1)/2
-			// If there are no short lived subnets, then the validator count is 0
-			if len(shortLived) == 0 {
-				currValidatorCount = 0
-			}
-			currAvgValidatorCount := ComputeNewAverage(prevAvgValidatorCount, prevNumObservations, currValidatorCount)
-
-			if err == sql.ErrNoRows {
-				// Insert new row
-				insertQuery := `
-				INSERT INTO validator_tracker (PeerID, ENR, Multiaddr, IP, Port, LastSeen, LastEpoch, PossibleValidator, AverageValidatorCount, NumObservations)
-				VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-				`
-				_, err = c.db.Exec(insertQuery, event.ID, event.ENR, event.Multiaddr, ip, port, event.Timestamp, event.Epoch, isValidator, currAvgValidatorCount, prevNumObservations+1)
-				if err != nil {
-					c.log.Error().Err(err).Msg("Error inserting row")
-				}
-				c.log.Trace().Str("PeerID", event.ID).Msg("Inserted new row")
-			} else if err != nil {
-				c.log.Error().Err(err).Msg("Error querying database")
-			} else {
-				// Update existing row
-				updateQuery := `
-				UPDATE validator_tracker
-				SET ENR = ?, Multiaddr = ?, IP = ?, Port = ?, LastSeen = ?, LastEpoch = ?, PossibleValidator = ?, AverageValidatorCount = ?, NumObservations = ?
-				WHERE PeerID = ?
-				`
-				_, err = c.db.Exec(updateQuery, event.ENR, event.Multiaddr, ip, port, event.Timestamp, event.Epoch, isValidator, currAvgValidatorCount, prevNumObservations+1, event.ID)
-				if err != nil {
-					c.log.Error().Err(err).Msg("Error updating row")
-				}
-				c.log.Trace().Str("PeerID", event.ID).Msg("Updated row")
-			}
-		default:
-			c.log.Debug().Msg("No validator metadata event")
-			time.Sleep(1 * time.Second) // Prevents busy waiting
-		}
 	}
 }
